requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
pydantic>=2.7.0

# LLMs / LangChain / LangGraph
langchain>=0.2.10
langgraph>=0.2.36
langchain-google-genai>=1.0.7
langchain-huggingface>=0.1.2
groq>=0.9.0
pymongo>=4.7.2

# Testing
pytest>=8.3.1
pytest-asyncio>=0.23.7
httpx>=0.27.0
mongomock>=4.1.2 